switch to JWT for authentication, and switch from md5 to sha512 for password hash

+++ client/user/full-name-unset.py
@@ -5,16 +5,18 @@
 import sys
 import os
 import requests
-from requests.auth import HTTPDigestAuth
+
 from pprint import pprint
 
 host = os.getenv("STARCAL_HOST", "127.0.0.1")
-email = os.getenv("STARCAL_EMAIL")
-password = os.getenv("STARCAL_PASSWORD")
+if not token:
+	print("Please set and export STARCAL_TOKEN")
+	sys.exit(1)
 
 r = requests.delete(
 	"http://%s:9001/user/full-name/" % host,
-	auth=HTTPDigestAuth(email, password),
+	headers={"Authorization": "bearer " + token},
+	#headers={'Authorization': 'bearer ' + },
 )
 print(r)
 try:
